pyproject.toml
[project]
name = "prism"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.14.0",
    "Pillow>=11.0.0",
    "python-dotenv>=1.1.0",
]
